multimodal_aifs/training/requirements.txt
# Minimal Requirements for Training
# Only the essential packages needed to run training

torch>=2.0.0
transformers>=4.30.0
deepspeed>=0.9.0
numpy>=1.20.0
PyYAML>=6.0
tqdm>=4.65.0
